feat: 원시값 포장 Number 클래스 추가
