+++ src/main/java/calculator/StringAddCalculator.java
@@ -5,11 +5,10 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class StringAddCalculator {
-    public static final int ZERO = 0;
-    public static final String REGEX_COMMA_OR_COLON_SEPARATOR = ",|:";
-    public static final String REGEX_CUSTOM_SEPARATOR = "//(.)\n(.*)";
-    public static final String ERROR_MESSAGE_NEGATIVE = "음수는 입력 불가능 합니다.";
-    public static final String ERROR_MESSAGE_ONLY_NUMBER = "숫자 이외의 값은 입력 불가능합니다.";
+    private static final int ZERO = 0;
+    private static final String REGEX_COMMA_OR_COLON_SEPARATOR = ",|:";
+    private static final String REGEX_CUSTOM_SEPARATOR = "//(.)\n(.*)";
+    private static final Pattern CUSTOM_PATTERN = Pattern.compile(REGEX_CUSTOM_SEPARATOR);
 
     private StringAddCalculator() {
     }
@@ -29,7 +28,7 @@ public class StringAddCalculator {
     }
 
     private static String[] splitText(String input) {
-        Matcher matcher = Pattern.compile(REGEX_CUSTOM_SEPARATOR).matcher(input);
+        Matcher matcher = CUSTOM_PATTERN.matcher(input);
         if (matcher.find()) {
             String customSeparator = matcher.group(1);
             return matcher.group(2).split(customSeparator);
@@ -39,17 +38,7 @@ public class StringAddCalculator {
 
     private static void validNumbers(String[] numbers) {
         for (String number : numbers) {
-            try {
-                validNegative(Integer.parseInt(number));
-            } catch (NumberFormatException e) {
-                throw new RuntimeException(ERROR_MESSAGE_ONLY_NUMBER);
-            }
-        }
-    }
-
-    private static void validNegative(int number) {
-        if (number < ZERO) {
-            throw new RuntimeException(ERROR_MESSAGE_NEGATIVE);
+            new Number(number);
         }
     }
